-moved constants.py from views package to main package
